Assignment1/Resume_Builder/requirements.txt
Flask==2.0.1
google-generativeai==0.8.1
python-dotenv==0.19.0
Werkzeug==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
pytest==6.2.5
reportlab==4.2.2
memory_profiler==0.61.0